// adapters/implementations/codex-app-server/src/protocol/generated/ServerRequest.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplyPatchApprovalParams } from './ApplyPatchApprovalParams';
import type { ExecCommandApprovalParams } from './ExecCommandApprovalParams';
import type { RequestId } from './RequestId';
import type { CommandExecutionRequestApprovalParams } from './v2/CommandExecutionRequestApprovalParams';
import type { FileChangeRequestApprovalParams } from './v2/FileChangeRequestApprovalParams';

/**
 * Request initiated from the server and sent to the client.
 */
export type ServerRequest =
  | { method: 'item/commandExecution/requestApproval'; id: RequestId; params: CommandExecutionRequestApprovalParams }
  | { method: 'item/fileChange/requestApproval'; id: RequestId; params: FileChangeRequestApprovalParams }
  | { method: 'applyPatchApproval'; id: RequestId; params: ApplyPatchApprovalParams }
  | { method: 'execCommandApproval'; id: RequestId; params: ExecCommandApprovalParams };